requirements.txt
apache-airflow==2.9.1
openai>=1.30.0
oci>=2.133.0
oracledb>=2.3.0
